app.py


from re import search

from click import prompt

import streamlit as st
from langchain_groq import ChatGroq
from langchain_community.utilities import ArxivAPIWrapper,WikipediaAPIWrapper
from langchain_community.tools import ArxivQueryRun,WikipediaQueryRun,DuckDuckGoSearchRun
from langchain.agents import initialize_agent,AgentType
from langchain.callbacks import StreamlitCallbackHandler


import os
from dotenv import load_dotenv

######Arxiv and wikipedia Tools

arxiv_wrapper=ArxivAPIWrapper(top_k_results=1,doc_content_chars_max=200)
arxiv=ArxivQueryRun(api_wrapper=arxiv_wrapper)



api_wrapper=WikipediaAPIWrapper(top_k_results=1,doc_content_chars_max=200)
Wiki=WikipediaQueryRun(api_wrapper=api_wrapper)


serach=DuckDuckGoSearchRun(name="Search")

st.title(" Langchain-Chat with search")




#####sidebar for settings

st.sidebar.title("Settings")
api_key=st.sidebar.text_inputs("Enter your Groq API Key:",type="password")


if "messages" not in st.session_state:
    st.session_state["messages"]=[
        {"role":"assistant","content":"Hi,I'm a chatbot who can search the web.How can i assist you today"}
    ]



for msg in st.session_state.messages:
    st.chat_messages(msg["role"]).write(msg['content'])


if prompt:st.chat_input(placeholder="What is machine learning?"):


st.session_state.messages.append({"role":"user","content":promt})



st.chat_messages("user").write(prompt)


llm=ChatGroq(groq_api_key=api_key,model_name="Llama3-8b-8192",streaming=True)

tools=[search,arxiv,Wiki]

search_agent=initialize_agent(tools,llm,agent=AgentType.Zero_SHOT_REACT_DESCRIPTION)


with st.chat_message("assistant"):
    st_cb=StreamlitCallbackHandler(st.container(),expand_new_thoughts=False)
    response=search_agent.run(st.session_state.messages,callbacks=[st_cb])

    st.session_state.messages.append({'role':'assistant',"content":response})

