requirement.txt
langchain-openai
langchain
python-dotenv
langchain_community
langchain_core
streamlit
ipykernel
pypdf
bs4
arxiv
pymupdf
wikipedia
langchain-text-splitters
chromadb
sentence-transformers
langchain_huggingface
faiss-cpu
duckdb
pandas
openai
langchain-groq
arxiv
wikipedia

mysql-connector-python
SQLAlchemy
validators==0.28.1

youtube_transcript_api

unstructured

